Implementation of data validation and accuracy
code: Modifications to the file classify_acess.py

Some new features were implemented in the algorithm. As:

- Validation of data;
- Acuraia of the model;
- Percentage of correctness of the model.

This model was very assertive close to 100% showing a tendency of the
result to be skewed.

For the next models check how the model could be improved by running
away from the data.